+++ src/browser.ts
@@ -1,65 +1,17 @@
-import assert from "node:assert/strict";
-
 import { BrowserContext, chromium, firefox, webkit } from "@playwright/test";
+import { BrowserType as PlaywrightBrowserType } from "@playwright/test";
 
 import { Logger } from "./logging.js";
-import {
-  BrowserType,
-  PersistentLaunchOptions,
-  Path,
-  RunConfig,
-} from "./types.js";
+import { BrowserType, PersistentLaunchOptions, RunConfig } from "./types.js";
 
-const browserTypeMapping = {
-  [BrowserType.Brave]: chromium,
-  [BrowserType.Chromium]: chromium,
-  [BrowserType.Gecko]: firefox,
-  [BrowserType.WebKit]: webkit,
-};
+type BrowserOptionsFunc = (config: RunConfig) => PersistentLaunchOptions;
+interface BrowserLaunchParams {
+  type: PlaywrightBrowserType;
+  options: BrowserOptionsFunc;
+}
+type BrowserTypeMapping = Record<BrowserType, BrowserLaunchParams>;
 
-const launchArgsEmpty = (): string[] => {
-  return [];
-};
-
-const launchArgsForBrave = (config: RunConfig): string[] => {
-  const launchArgs = launchArgsForChromium(config);
-  launchArgs.push("--disable-brave-update");
-  launchArgs.push("--disable-sync");
-  return launchArgs;
-};
-
-const launchArgsForChromium = (config: RunConfig): string[] => {
-  const launchArgs = ["--disable-features=MacAppCodeSignClone"];
-  if (config.profile) {
-    launchArgs.push(`--profile-directory="${config.profile}`);
-  }
-  return launchArgs;
-};
-
-const launchArgsForGecko = (): string[] => {
-  return launchArgsEmpty();
-};
-
-const launchArgsForConfig = (config: RunConfig): PersistentLaunchOptions => {
-  let launchArgsFunc;
-
-  switch (config.browser) {
-    case BrowserType.Brave:
-      launchArgsFunc = launchArgsForBrave;
-      break;
-    case BrowserType.Chromium:
-      launchArgsFunc = launchArgsForChromium;
-      break;
-    case BrowserType.Gecko:
-      launchArgsFunc = launchArgsForGecko;
-      break;
-    case BrowserType.WebKit:
-      launchArgsFunc = launchArgsEmpty;
-      break;
-  }
-  assert(launchArgsFunc);
-  const browserArgs = launchArgsFunc(config);
-
+const launchOptionsDefault = (config: RunConfig): PersistentLaunchOptions => {
   // We implement *not* preserving tabs and pages that are open in the user
   // browser state by
   // 1. launching the browser in offline mode
@@ -67,10 +19,8 @@ const launchArgsForConfig = (config: RunConfig): PersistentLaunchOptions => {
   //    user-data-dir/profile state
   // 3. and then, enabling networking.
   const startInOfflineMode = !config.preservePages;
-
-  const defaultBrowserArgs: string[] = [];
-  const launchOptions: PersistentLaunchOptions = {
-    args: defaultBrowserArgs.concat(browserArgs),
+  return {
+    args: [],
     executablePath: config.binary,
     headless: false,
     offline: startInOfflineMode,
@@ -81,28 +31,63 @@ const launchArgsForConfig = (config: RunConfig): PersistentLaunchOptions => {
     serviceWorkers: "block",
     timeout: config.timeout * 1000,
   };
-
-  return launchOptions;
 };
 
-const getContext = async (
-  browser: BrowserType,
-  userDataDir: Path,
-  launchArgs: PersistentLaunchOptions,
-): Promise<BrowserContext> => {
-  const browserType = browserTypeMapping[browser];
-  assert(browserType);
-  return await browserType.launchPersistentContext(userDataDir, launchArgs);
+const launchOptionsBrave = (config: RunConfig): PersistentLaunchOptions => {
+  const options = launchOptionsChromium(config);
+  options.args.push("--disable-brave-update");
+  options.args.push("--disable-sync");
+  return options;
+};
+
+const launchOptionsChromium = (config: RunConfig): PersistentLaunchOptions => {
+  const options = launchOptionsDefault(config);
+  options.args.push("--disable-features=MacAppCodeSignClone");
+  if (config.profile) {
+    options.args.push(`--profile-directory="${config.profile}`);
+  }
+  return options;
+};
+
+const launchOptionsGecko = (config: RunConfig): PersistentLaunchOptions => {
+  return launchOptionsDefault(config);
+};
+
+const launchOptionsWebKit = (config: RunConfig): PersistentLaunchOptions => {
+  return launchOptionsDefault(config);
+};
+
+const browserTypeConfigMapping: BrowserTypeMapping = {
+  [BrowserType.Brave]: {
+    type: chromium,
+    options: launchOptionsBrave,
+  },
+  [BrowserType.Chromium]: {
+    type: chromium,
+    options: launchOptionsChromium,
+  },
+  [BrowserType.Gecko]: {
+    type: firefox,
+    options: launchOptionsGecko,
+  },
+  [BrowserType.WebKit]: {
+    type: webkit,
+    options: launchOptionsWebKit,
+  },
 };
 
 export const launch = async (
   logger: Logger,
   config: RunConfig,
 ): Promise<BrowserContext> => {
-  const browserType = config.browser;
-  const userDataDir = config.userDataDir;
-  const launchArgs = launchArgsForConfig(config);
+  const paramsForBrowser = browserTypeConfigMapping[config.browser];
 
-  logger.info("Launching browser with args: ", { ...launchArgs, userDataDir });
-  return await getContext(browserType, userDataDir, launchArgs);
+  const launchOptionsForConfigFunc = paramsForBrowser.options;
+  const opts = launchOptionsForConfigFunc(config);
+
+  const browser = paramsForBrowser.type;
+  const userDataDir = config.userDataDir;
+  logger.info("Launching with options: ", { ...opts, userDataDir });
+  const context = await browser.launchPersistentContext(userDataDir, opts);
+  return context;
 };